Platform set for SKR v1.4 Turbo LPC1769
Setup to support an Ender 3 with a E3D Hemera hotend
Motor Driver TMC2208 UART Mode.

The FAN0 connector doesn't have enough current to power
the larger part cooling fan. I have had to move it to the
HE1 (2nd hotend heater) which is used in default config
for the hotend fan.